+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-30 10:34:56
-// Commit: fix(state): prevent double tap duplicate action in state
+// Updated at: 2026-05-30 12:54:15
+// Commit: fix(state): fix memory leak by cancelling state subscription
 
 class stateComponentState {
-  final String commitRef = '1609';
-  final String lastMessage = 'fix(state): prevent double tap duplicate action in state';
-  final DateTime timestamp = DateTime.parse('2026-05-30T10:34:56+05:00');
+  final String commitRef = '1616';
+  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-30T12:54:15+05:00');
 }